refactor(entities): Changed entity type files => entity.types.ts
